Tpetra: Manually fixing all of the stuff opencode screwed up in this refactor
